03_ComputerVision/lab/gan_exercise_1/requirements.txt

matplotlib==3.8.2
ema-pytorch==0.3.1
